spinaltap-common: Update metrics for the future
